Add tox for testing
This commit adds to [dev] section the tox library so as
to automate testing workflows. Since tox will create the
required environment to run poetry within, but poetry is
commonly found as a system-wide asset, we tell tox to
use that by whitelisting poetry as a required, but external,
asset.

Moreover, tox has been marked as a dev dep of the project
so easy testing/development workflow so as to keep them in
the same virtual environment