Include errors in finger results
In order to better catch errors during the scraping stage, exception
handling has been improved to the point of self-implementing a base
exception for all the project: WpokeException. This type inherits from
'BaseException', the recommended type to be overriden should error handling
behaviour to be customized.

This commit also adds an 'errors' field to FingerResult type to store which
errors have been yielded for every finger.