refactor: optimize disccount calculation
This commit changes how disscounts are calculated. Previous to this
commit, a for-loop was being used to iterate over all basket items
so as to check which ones disccounts could be applied to. As baskets
use a map internally, these changes focuse on the fact that  complexity
can be reduced from O(n) to O(1).

In the first iteration of this project, baskets used a slice instead of
a map internally. Then, disccounts systems were implemented. Afterwards,
baskets were altered to use employ a hashmap. However, as tests were
still passing since slices and maps employ the same "range" interface, I
forgot to change that logic. Maybe how to access items from a basket
should be an implementation detail kept within the basket object itself,
but that's another commit story.